- hopefully fixed the concurrerModificationError once and for all - reused "YggdRamusSwitchSkill" and turned it into ItemButtonInteractionToServer - made some more changes to the Asgard Fractal for upcoming abilities for that weapon

+++ src/main/java/yerova/botanicpledge/common/entitites/projectiles/AsgardBladeEntity.java
@@ -48,7 +48,7 @@ public class AsgardBladeEntity extends EntityProjectileBase {
 
     }
 
-    public AsgardBladeEntity(Level world, LivingEntity thrower, LivingEntity target, int damage) {
+    public AsgardBladeEntity(Level world, LivingEntity thrower, LivingEntity target, float damage) {
         super(BPEntities.ASGARD_BLADE.get(), world, thrower);
         setThrower(thrower);
         setTargetPos(new BlockPos(target.getX(), target.getY(), target.getZ()));
